Day27

+++ java-basic/app/src/main/java/com/eomcs/oop/ex02/step14/App.java
@@ -1,13 +1,16 @@
 package com.eomcs.oop.ex02.step14;
 
-// 데이터 타입 정의
-// 0) 낱개 변수 사용
-// 1) 성적 데이터를 저장할 메모리를 새로 정의:
-// - '사용자 정의 데이터 타입(user-defined data type)'이라 부른다.
-// - 여러개로 이루어진 한 묶음으로 다루면 관리하기가 쉽다.
-// 2) 리팩토링: 메서드 추출(extract method)
-// 3) 리팩토링: 메서드 추출(extract method) 한 개의 메서드는 한 개의 기능을 수행해야 한다.
-
+// 데이터 타입 정의  
+// 1) 클래스 사용전: 
+//    - 낱개 변수 사용
+// 2) 성적 데이터를 저장할 메모리를 새로 정의: 
+//    - '사용자 정의 데이터 타입(user-defined data type)'이라 부른다.
+//    - 여러 개로 이루어진 데이터를 한 묶음으로 다루면 관리하기가 쉽다.
+// 3) 리팩토링: 
+//   - 메서드 추출(extract method): 중복되는 코드가 있으면 별도의 메서드로 분리한다.
+// 4) 리팩토링: 
+//   - 메서드 추출(extract method): 한 개의 메서드는 한 개의 기능을 수행해야 한다.
+//
 public class App {
   public static void main(String[] args) {
 
@@ -36,17 +39,20 @@ public class App {
     printScore(s1);
     printScore(s2);
     printScore(s3);
+
   }
 
-  // printScore() 출력 기능만 담당하도록 변경한다.
+
+  // printScore()는 출력 기능만 담당하도록 변경한다.
   static void printScore(Score s) {
-    System.out.printf("%s: %d, %d, %d, %d, %.1f\n", s.name, s.kor, s.eng, s.math, s.sum, s.aver);
+    System.out.printf("%s: %d, %d, %d, %d, %.1f\n", 
+        s.name, s.kor, s.eng, s.math, s.sum, s.aver);
   }
 
-  // compute()
+  // 합계, 평균은 별도에 메서드에서 처리하게 한다.
   static void compute(Score s) {
     s.sum = s.kor + s.eng + s.math;
-    s.aver = (float) s.sum / 3;
+    s.aver = s.sum / 3f;
   }
 }
