+++ java-basic/app/src/main/java/com/eomcs/oop/ex02/step17/vo/Score.java
@@ -1,13 +1,11 @@
 package com.eomcs.oop.ex02.step17.vo;
 
-// ## 사용자 정의 데이터 타입 만들기
-// - 특정 유형의 데이터를 담을 전용 메모리(변수)를 설계한다.
-// - 학생 데이터를 묶음으로 다룰 수 있도록 변수 그룹을 정의한다.
+// 다른 패키지에서도 이 클래스를 사용할 수 있도록 하려면
+// public 으로 공개해야 한다.
 public class Score {
-  // 인스턴스 변수(instance variable; instance field)
-  // - new 명령으로 생성되는 변수이다.
-  // - 데이터를 개별적으로 다루고 싶을 때 인스턴스 변수로 선언한다.
-  //
+
+  // 다른 패키지에서 이 설계도에 따라 만든 변수에 접근할 수 있도록 하려면
+  // 접근 범위를 public 으로 공개해야 한다.
   public String name;
   public int kor;
   public int eng;
@@ -15,9 +13,9 @@ public class Score {
   public int sum;
   public float aver;
 
+  // 다른 패키지에서 메서드를 사용할 수 있도록 하려면 public 으로 공개해야 한다.
   public void compute() {
     this.sum = this.kor + this.eng + this.math;
-    this.aver = (float) this.sum / 3;
-
+    this.aver = this.sum / 3f;
   }
 }
